+++ src/View/StateViewInit.java
@@ -37,7 +37,7 @@ public class StateViewInit extends StateBasedGame implements Runnable {
 	 * @param resWidth
 	 * @param resHeight
 	 */
-	public StateViewInit(String title, boolean fullscreen, boolean grabMouse, int targetFramerate, int resWidth, int resHeight, float scaleX, float scaleY){
+	public StateViewInit(String title, boolean fullscreen, boolean grabMouse, int targetFramerate, int resWidth, int resHeight, double scaleX, double scaleY){
 		super(title);
 
 		this.view = new View(PLAY_STATE, scaleX, scaleY);
@@ -103,7 +103,7 @@ public class StateViewInit extends StateBasedGame implements Runnable {
 		}
 	}
 
-	public void setRenderPoint(float x, float y){
+	public void setRenderPoint(double x, double y){
 		view.setRenderPoint(x, y);
 	}
